Upgrade codecov-action to v2 in CI workflow
The `codecov-action` used in the GitHub Actions CI workflow has been upgraded from v1 to v2. This was done to keep up with the latest changes and recommendations by `codecov-action`. Note that in v2, it automatically picks up the coverage reports, so there's no need for the `with` parameter anymore.